+++ blog-main-server/ChangeLogs.md
@@ -1,0 +1,25 @@
+## Change Logs
+相关文章：[微服务项目系列文章](http://www.veryitman.com/2019/05/28/%E5%BE%AE%E6%9C%8D%E5%8A%A1%E9%A1%B9%E7%9B%AE%E7%B3%BB%E5%88%97%E6%96%87%E7%AB%A0/)
+
+### 2020.12.25
+* 重构登录和注册；
+
+### 2020.12.02
+* 集成 Redis;
+* 相关文章：[微服务: SpringBoot 集成 Redis](http://www.veryitman.com/2020/12/05/%E5%BE%AE%E6%9C%8D%E5%8A%A1-SpringBoot-%E9%9B%86%E6%88%90-Redis/)
+
+### 2020.02-09
+* 更新相关博客并修复代码bug；
+* 重构相关逻辑；
+
+### 2019.11.10
+* 集成 `MySQL`;
+* MySQL 安装和使用；
+
+### 2019.09.15
+* 实现登录和注册；
+
+### 2019.05-08
+* 决心写 `Springboot` 相关文章，并以实例为主；
+* 搭建服务；
+* 基础建设，入门知识整理；
