+++ blog-main-server/ms-user/pom.xml
@@ -68,7 +68,6 @@
             <optional>true</optional>
         </dependency>
 
-
         <!--jsr-->
 
         <dependency>
