+++ blog-main-server/ms-user/src/main/java/com/veryitman/user/controller/MSSignupController.java
@@ -4,10 +4,7 @@
 package com.veryitman.user.controller;
 
 import com.veryitman.core.model.MSResponse;
-import com.veryitman.user.model.MSUser;
-import com.veryitman.user.model.MSUserResponseEnum;
-import com.veryitman.user.service.MSUserService;
-import com.veryitman.user.util.MSUserUtil;
+import com.veryitman.user.service.MSUserSignupService;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiImplicitParam;
 import io.swagger.annotations.ApiImplicitParams;
@@ -15,16 +12,17 @@ import io.swagger.annotations.ApiOperation;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.List;
-import java.util.Map;
-
 @Api(value = "signup", tags = "用户模块-注册")
 @RestController
 @RequestMapping(value = "signup") // 注意这里不要在signup前后加"/"
 public class MSSignupController {
 
+    private MSUserSignupService userSignupService;
+
     @Autowired
-    private MSUserService userService;
+    public void setUserSignupService(MSUserSignupService userSignupService) {
+        this.userSignupService = userSignupService;
+    }
 
     @CrossOrigin(origins = {"*"})
     @PostMapping(value = "/name")
@@ -34,33 +32,7 @@ public class MSSignupController {
             @ApiImplicitParam(name = "userpwd", value = "注册的密码", required = true)
     })
     public MSResponse signup(@RequestParam(value = "username") String userName, @RequestParam(value = "userpwd") String userPwd) {
-        MSResponse response = new MSResponse();
-        MSUser user = null;
-        if (null == userName || null == userPwd || userName.length() <= 0 || userPwd.length() <= 0) {
-            MSUserResponseEnum signupError = MSUserResponseEnum.SignupInvalidInfo;
-            response.setMsg(signupError.getMsg());
-            response.setCode(signupError.getCode());
-        } else {
-            // 创建user表
-            userService.createUserTable();
-            // 检查用户数据库的‘user’表中是否有该用户？
-            List<Map> query_users = userService.queryUserByUserName(userName);
-            if (null == query_users || query_users.isEmpty()) {// 没有该用户的数据
-                user = MSUserUtil.createUser(userName, userPwd);
-                // 插入一条用户数据到数据表中
-                userService.addUser(user);
-                response.setCode(MSUserResponseEnum.SUCCESS.getCode());
-                response.setMsg(MSUserResponseEnum.SUCCESS.getMsg());
-            } else {// 用户数据库的‘user’表中有该用户信息
-                // 返回错误信息：该用户已经注册过了
-                MSUserResponseEnum signupError = MSUserResponseEnum.SignupHasExistUser;
-                response.setMsg(signupError.getMsg());
-                response.setCode(signupError.getCode());
-            }
-        }
-
-        response.setResults(user);
-
+        MSResponse response = userSignupService.signupUsingUserName(userName, userPwd);
         return response;
     }
 }
